[Web3v1] refactor codegen structure
