+++ packages/target-web3-v1/package.json
@@ -19,7 +19,7 @@
     "lint:fix": "yarn lint --fix",
     "typecheck": "tsc --noEmit --incremental false --composite false && tsc --noEmit --incremental false --composite false -p tsconfig.types.json",
     "clean": "rm -rf dist && rm -f tsconfig.build.tsbuildinfo",
-    "post-build": "cp ./src/types.d.ts ./dist/",
+    "post-build": "cp -rf ./src/static ./dist/",
     "generate-types": "../typechain/dist/cli/cli.js --target=`pwd`/dist/index.js --outDir ./types/ '../../contracts/compiled/*.abi'",
     "test": "yarn generate-types && mocha --config ../../.mocharc.js",
     "test:fix": "yarn lint:fix && yarn format:fix && yarn test && yarn typecheck"
