+++ packages/target-web3-v1/src/index.ts
@@ -4,7 +4,7 @@ import { readFileSync } from 'fs'
 import { TsGeneratorPlugin, TContext, TFileDesc } from 'ts-generator'
 import { extractAbi, parse, getFilename } from 'typechain'
 
-import { codegen } from './generation'
+import { codegen } from './codegen'
 
 export interface IWeb3Cfg {
   outDir?: string
@@ -46,7 +46,7 @@ export default class Web3V1 extends TsGeneratorPlugin {
     return [
       {
         path: join(this.outDirAbs, 'types.d.ts'),
-        contents: readFileSync(join(__dirname, './types.d.ts'), 'utf-8'),
+        contents: readFileSync(join(__dirname, './static/types.d.ts'), 'utf-8'),
       },
     ]
   }
